move translations into assets folder
